adding dotenv and session

+++ web-form-db/src/app.js
@@ -1,7 +1,7 @@
 const express = require('express');
-const fs = require('fs');
 const path = require('path');
 const bodyParser = require('body-parser');
+const session = require('express-session');
 
 const dbService = require('./db-service.js')
 
@@ -9,9 +9,15 @@ const app = express();
 const PORT = 3002;
 
 // Middleware
+// injects the urlencoded data into request.body
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 app.use(express.static(path.join(__dirname, 'html')));
+app.use(session({
+    secret: 'my-secret-key',
+    resave: true,
+    saveUninitialized: true
+}));
 
 // Routes
 app.get('/', (req, res) => {
@@ -21,11 +27,10 @@ app.get('/', (req, res) => {
 // Login endpoint
 app.post('/login', (req, res) => {
     const { username, password } = req.body;
-    // You can implement your login logic here, like checking credentials against a database
-    // For simplicity, I'm just checking if username and password are both 'admin'
 
-    dbService.valicateUser(username, password, (isValid) => { 
+    dbService.validateUser(username, password, (isValid) => { 
         if (isValid) {
+            req.session.authenticated = true;
             res.redirect('/students');
         } else {
             res.redirect('/login-fail');
@@ -36,6 +41,10 @@ app.post('/login', (req, res) => {
 // Endpoint to display list of students
 app.get('/data', (req, res) => {
 
+    if (!req.session.authenticated) {
+        res.redirect('/');
+        return;
+    }
     dbService.getStudents((students) => { 
         if (!students) {
             students = '[]';
@@ -47,11 +56,19 @@ app.get('/data', (req, res) => {
 
 // List page
 app.get('/students', (req, res) => {
+    if (!req.session.authenticated) {
+        res.redirect('/');
+        return;
+    }
     res.sendFile(path.join(__dirname, 'html', 'students.html'));
 });
 
 // Registration page
 app.get('/register', (req, res) => {
+    if (!req.session.authenticated) {
+        res.redirect('/');
+        return;
+    }
     res.sendFile(path.join(__dirname, 'html', 'register.html'));
 });
 
@@ -61,7 +78,10 @@ app.get('/login-fail', (req, res) => {
 
 // Registration endpoint
 app.post('/register', (req, res) => {
-
+    if (!req.session.authenticated) {
+        res.redirect('/');
+        return;
+    }
     const student = { firstName, lastName, email, dob, gender, address } = req.body;
 
     dbService.saveStudent(student, (isValid) => { 
@@ -76,6 +96,10 @@ app.post('/register', (req, res) => {
 // Delete studenta student record by ID
 app.delete('/students/:id', (req, res) => {
     const id = parseInt(req.params.id);
+    if (!req.session.authenticated) {
+        res.redirect('/');
+        return;
+    }
 
     dbService.deleteStudent(id, (isValid) => { 
         res.redirect('/students');
